[IMP] core: remove sql_log attribute
This thing is a pain in the ass, it requires setting the logger
correctly *then* remembering that you also have to set and re-set this
flag on the cursor (when trying to log queries within a specific
scope).

It's a complete waste of time, `isEnabledFor` is unlikely to take
ages given the logger sits very shallowly (at worst `logging` might
have to go up 2 levels to the root logger before it finds the
effective level) so it's not like there's any sort of humongous perf
hit (especially compared to the SQL query itself).

Make `sql_log` a computed property based on the current log level,
this way tests just need to update the loglevel of `odoo.sql_db` for
things to work as they should, warn (and ignore) on setting it.