[FIX] crm: Fix lead multicompany issue
Context
-------

Lead get assigned a different company than the one defined on the
salesman after being merge with another lead.

This happen because some lead assigned to a user_id have no company set
are merge with lead that receive self.env.company during automated
assignation to the team and than get merged with the lead without
company. So the company of the new lead is written on the old one.

Behavior before this PR
-----------------------

When the crm.team have the field company_id = False
With the following flow

 1) Create a lead without a user_id and a team_id
 2) Assign a team to the lead
 3) Assign a user_id

The status of the company field on the lead is the following

 1) set the company of the env.user
 2) Keep the company of the lead
 3) set the user company if the current company is not one of the allowed company of the user

Behavior after this PR
----------------------

 1) set the company of the env.user
 2) set the company of the team even if it's False
    (so erase the company if the team has no company set)
 3) set the user company if the current company is not one of the allowed company of the user