[IMP] base, website: add bulk delete action on ir.ui.view
Before this commit, there was no easy way to delete a whole view tree
hierarchy, you would need to somehow get all the active and inactive views from
that hierarchy in the list view (which is basically impossible), and then
select them all to unlink them.

This commit adds a "Bulk Delete" entry in the control panel action entries in
both the form and list view.
It also adds a "Bulk Delete" button to the QWeb view (hierarchy view).

This is particularly useful in multi-website environment when you need to
delete a whole view hierarchy that had been cow'd.