[IMP] survey: add image zoom widget and name
Before this commit, images displayed on survey question answers were small
and hard to read, and nothing existed in order to zoom on them properly.

Also, in backend, only the image was present in columns of question answer
list view. This was not helpful to match a line to an image, as default list
view for binary files shows but the  size of the image.

AFTER THIS COMMIT

When clicking on such an image, the image is zoomed on and displayed in a
foreground modal. The image ratio is kept and size emphasized. The max_height
and max_width of the images are doubled to 512 px to compromise between image
resolution and space. A hovering frame is visible on the images, as well as
a zoom in cursor, to indicate we can click on them to zoom in. The rest of the
answer frame is the selection area.

Clicking anywhere on the modal closes it.

Also, on backend, add a column in tree view to keep the image name after
loading the image file. This is editable and can be used as an additional
information on image in backend (version, ...)

Task-2566584