[ADD] base: add a websocket controller to the base module
A Websocket controller has been added in the base module. It will
handle incoming websocket connections ('/websocket') and proceed
to the connection upgrade if the handshake is successfull.

Once the connection is established, the controller will wait for
incoming frames and dispatch them to the appropriate route according
to the "path" that have been given in the payload. Optionnally, keyword
arguments can be passed with the "kwargs" key. An example of a correct websocket
request could be:

websocket.send(json.stringify({
    "path": "/echo",
    "kwargs": {
        "message": "hello odoo websockets"
    }
}));

Websockets route are very much similar to http/json ones, only the type
changes. To follow the previous example, here is a user route that will
echo the message back to the client. The value which is returned by the
function will be send through the socket.

@​​route('/echo', type='websocket', auth='user')
def echo(self, message):
    return message

A WebsocketRequest object is made available (odoo.websocket.ws_request).
This request hold a reference to the: websocket, env, session, db, registry.

Two lifecycle hooks are available: Websocket.onopen, Websocket.onclose.
Those functions will be call respectivly after the websocket opening
and after the websocket closure. Both of those will be passed the websocket
object.

The test_websockets module has been added in order to test the websocket
routing/handshake.