[IMP] mail: add a post server action
Currently only an email action based on template exists in server actions when
having mail app installed. It basically sends an email based on a mail template.

However being able to post a message on record is also useful. Instead of
sending emails it post on a document as a comment. Notification flow is the
classic of from post: followers, specified partners, Inbox/Email, ...

Task-2613245

Closes odoo/odoo#45640