[IMP] mail: add author field on mail template
When you configure a mail template, you can specify an author for
the mail. This allows to simplify the setting of a template because
you no longer need to know the format to select the email of the author.
Example: for a quotation, you can set the author to "{{object.user_id.id}}"
and the email_from field would be set automaticaly to the email of the
saleperson if the email from is not specified.

Additionaly, when an user sending the message is different from the author,
we add the sender name next to the name of the author with "impersonated by xxx"
in the chatter.

task-2091859
COM PR: odoo/odoo#42624