[IMP] web,web_editor,mass_mailing: new dynamic placeholder
Introduce a new Dynamic Placeholder Generator
in mass mailing and SMS creation.

Using a specific character (#) in subject or body will trigger
the opening of the field selector popup, which can be use to select
the wanted field and enter a default value. on validation the dynamic
placeholder will be inserted in the corresponding input/editable.

task-2154624