[IMP] fleet, hr_recruitment: improve the document flow for fleet vehicle.
Purpose of the task is to add stat button on vehicle form view
for attachment and on click of that button open the attachment
kanban view. from this button user can create an attachment.

Also some code refactor for hr_recruitment. removed unused field
document_ids and rename the count field as attachment_count instead
of document_count(due to conflict with name in document.mixin).

closes odoo/odoo#78592
task-2144780