[IMP] mail: disable upload in oe_chatter
Currently, we can not disable the 'add an attachment' button in chatter.

Purpose of the task is to disable the upload button so user can upload the
document using stat button of document app.

So in this commit, Added the options 'has_attachment_upload' chatter field
so user can  disable the upload from bottom chatter of the form view.
For ex:
<field name="message_ids/message_follower_ids"
options="{'has_attachment_upload': [('documents_fleet_settings', '=', True)]}"/>

closes odoo/odoo#78592
task-2144780