[WIP] survey: allow to resolve question dependence on chain creation
Due to the behavior of X2many fields, the questions created when editing
a survey are not usable immediately after creation to be conditional
triggers for other questions. The parent survey has to be saved for the
questions to be available. This is not a smooth experience for the user.

This commit will impose auto saving of the parent survey at each new
question added, and adds a _compute attribute for its child questions
(similarly to #6675).
This means that the questions are also saved right after being introduced,
allowing to immediately use the field validation possibilities.

Note that to be able to immediately validate the question "validation"
fields and use new questions as triggers for the next ones, all rpc
calls are now done synchronously.

It also allows to activate scoring and add scores to question/answers
without requiring the extra steps of saving and re-entering edit mode
for the survey.

Task-2623482