http.py _serve_static
Hello reviewers, here are a few implementations insights:

The `send_file` utility have been split in two: `send_filepath` that
takes a path-like and `send_file` that takes either a file-objet or
the content right away (as str/bytes). This new `send_filepath` is a
request from the Odoo SH team and should be the prefered way to stream
a file located on the filesystem.

The original function has been deprecated to be moved on Request, Antony
prefers it this way, I personally don't really care.

The `addons_manifest` is a cache for module manifest, it is non-sense
it is located in the http module and I would like to move it elsewhere,
that's duty for another PR. Because it is populated upon first request
in `load_static` we couldn't define `Application.statics` as a lazy
property (which makes me and rco very sad).

You notice that `_serve_static` raises `werkzeug.exceptions.NotFound` in
the case the requested file couldn't be served. You may wonder how we
handle exceptions, it in a future commit.