http.py http_dispatch
Before this refactor, the code of `Response.load` was inlined in
`route_wrapper`. It was moved to a dedicated method in order to document
it.

Note that `Response.load` is called by `route_wrapper` instead of
`http_dispatch`. When a http route with is override by another
controller, it is required that super() returns a Response already.