base ir_http
The model have been refactored to use the updated http.py. The main
contribution is that `http.py` is now responsible to call the four
`_match()`, `_authenticate()`, `_pre_dispatch()` and `_dispatch()`.

Each method have been updated, the `request.uid`, `request.context` and
alike magic properties have been replaced by the more explicit
`request.update_env` and `request.update_context`.

The original environment is created using the uid and context stored on
the session right away. It uses `None` and a generic context when the
session is missing. The `_authenticate()` method and its
`_auth_method_...()` compagnions have been refactored accordingly.

The `_pre_dispatch` method is a new one, here we have a proper example
to better describe its usage. When one request a db-compatible endpoint,
it is possible the route contains a `<model(model):id>` dynamic argument,
when a user request this URL he must provide a record id as an integer.
Remplacing this integer by the real record should be done prior to
dispatch the request to the route. This is what `_pre_dispatch()` is
made for, to prepare the system to deal with the current request. Note
that because `_pre_dispatch()` is called prior to
`http._<type>_dispatch()`, the because body have not been parsed yet
thus `request.params` is not ready and should not be used.

What is nice with `_pre_dispatch()` is that it doesn't show up in the
traceback if there's an error in the controller. Shorter tracebacks on
errors is also a feature of this refactor.