http_routing and website madness READ MY MESSAGE
This one is very though. I highly suggest you first read this message
first. I'm going to explain what was done in `http_routing`/`website`
before, how it clashes with this refactor and how we managed to make the
two modules compatible with the new design.

First and foremost, `http_routing` is a technical module that aim to
provide the minimum viable compatibility code between portal and
website. Its primary job is to take care of the lang inserted in the
path of URLs, e.g. `fr` in `/fr/mon_blog`. Both when routing a request
with a lang in the URL and when rendering templates with multilang
support.

Next to `http_routing` is website, the module used by customers to
create pretty web page accessible online. Website uses a different
routing logic than the backend, webpages are registered in routing map
of werkzeug but instead delivered by website dedicated code. It means
that **all** request targeting a website page thrown at the werkzeug
router **fail** with a HTTP 404 error. The reality is that website
abuses the fallback mechanism (`_serve_fallback`) to deliver its pages.

Using `_serve_fallback` as the standard way to deliver pages is broken
by design. It is the least crappy way to deliver content as long as
website page don't have a dedicated path prefix. Using a path prefix
(e.g. `p` in `/p/fr/mon_blog`) it would have been possible to route the
request to a dedicated endpoint using the same router as the backend and
with no change to the HTTP dispatching code. Sadly, the business does
not want such prefix so we have to stick with a broken design.

It is broken because prior to serving a page, website needs to setup A
LOT of stuff on the system. It needs to ensure a proper user is set on
the environment, it needs to setup the GeoIP database and it also needs
to determine the lang the user requested the page. Due to the very
fragile implementation, is was necessary to differentiate requests
delivered throught website from the one delivered via the backend. It
was decided to save a magical attribute on the `request` object:
`is_frontend`. Not ... [TODO]

[TODO]

Before this refactor, the solution implemented in `http_routing` was to
take over the entire request dispatching. As the method `_dispatch` in
`base/ir.http` was responsible to call the other `_match`,
`_authenticate` and `WebRequest._dispatch` methods, they override it to
implement their own

Before this commit, the flow was as follow:

	App._dispatch
	-> ir_http._dispatch: call match, authenticate and dispatch
	-> (Json|Http)Request._dispatch
	-> endpoint

With this commit, it is:

	App.__call__
	-> Req._serve_db
	-> Req._serve_ir_http: call match, authenticate and dispatch
	-> Req._(http|json)_dispatch
	-> ir_http._dispatch
	-> endpoint

One important change considering http_routing is that ir_http._dispatch
does **no longer** control the flow match/authenticate/dispatch thus it
is **no longer** possible to override a single method to *take over* the
entire dispatching.