[FIX] web_editor: not deduce selected color from unselected element
Since [1] selecting a part of a text to apply a custom gradient right
away does consider the end of the text node that precedes the newly
created `<font>` node as the begin of the selected text when deducing
what is the current selection's color. Because of this, instead of the
newly applied gradient, that previous text's color is used to populate
the color palette - therefore closing the gradient editor which was just
opened.

This commit changes the was the selected color is determined:
Before this commit if a selection started at the end of an element it
was still used to deduce the selected color.
After this commit if a selection starts at the end of the start element,
the selected color is deduced from the next leaf of the start element
instead.

Steps to reproduce:
- drop a "Text" snippet
- select a word (e.g. "general" in the second paragraph) either through
click and move or by double click
- click on the font color toolbar icon
- click on the "Gradient" tab of the color palette
- click on the "Custom" button
=> the gradient editor appeared but vanished immediately
(to make it work, the text had to be re-selected)

[1]: https://github.com/odoo/odoo/commit/9cb5a138244783580c2d1baa4dc01d1824ed20f4

task-2666200