[FIX] pos_coupon: free product is not properly added in the order
For free product promo (or coupon) program, there are 2 scenarios:
1. reward product = criterion product
2. reward product != criterion product

**Scenario 1**

For 1, there is no problem and it behaves like this:

Promo: Buy 3 Beers, Get 1 Beer

After adding 3 beers, no reward is automatically added yet. The order looks
like this:
```
 3 Beer
```
To get the free beer, the user has to add another beer in the order. So,
after clicking again the Beer product, the order becomes:
```
 4 Beer
-1 Free Product - Beer
```
Let's call the -1 line as virtual reward line. Its role is to negate the
effect of the 4th beer.

**Scenario 2**

The bug is in this scenario.

Example Promo: Buy 4 Notebooks, Get 1 Pen

Bug: After adding 4 Notebooks, there is no reward line, but when the order
became 5 Notebooks, -1 Free Product - Pen is added. Here's how it looks like:
```
 5 Notebook
-1 Free Product - Pen
```
Fix: After adding 4 Notebooks, 1 Pen should be automatically added, together
with it's virtual reward line (the one that will negate the cost of the
free Pen) like so:
```
 3 Notebook
 1 Fancy Notebook <- the 4th Notebook
 1 Pen
-1 Free Product - Pen
```

TASK-ID: 2692252