[IMP] mail: Filter attachment in oe_chatter
Currently, we can not filter attachments in chatter.

This feature will prevent the document uploaded from the document app from appearing
in the chatter.

Purpose of the task is to filter the chatter attachments so the user can upload the
document using stat button of document app.

So in this commit, Added the options 'has_visible_document' in the chatter field
so the user can filter the attachment of the form view.

For ex:
<field name="message_ids/message_follower_ids"
options="{'has_visible_document': True/>

closes odoo/odoo#78592
task: 2144780