[FIX] web: do not display notebook ul if thee are no pages
Before this commit: when there is no page in notebook, only notebook
tag is defined in xml, form still displays notebook header's ul because of
that blank grey line is displayed.

After this commit: if there is no page to display in notebook then header's
ul is not displayed and hence blank line is not displayed in form view.

task-2653806