[TMP] sale_coupon: deactivate coupon post
``_send_reward_coupon_mail`` tries to post on a sale order a template rendered
on coupon model. This results in actually calling ``_message_notify`` on
MailThread and redirecting by change to sale order using model / res_id.

This is going to fail as we are going to be stricter on what API offers notably
on model/res_id manipulation, as well as limiting more cleanly post API to
mail thread objects.

This is temporarily deactivated until a solution is found after cleaning the
composer and its onchanges.