[IMP] website_sale: improve design of templates
This commit improves the design of dynamic products templates.
It introduces templates using multiples rows and fine tunes the
design of the other ones.
To make thumbnails work in the template selector, an extra
data-attribute can now be processed by the snippet option.

- data-thumb: it references the location of the thumbnail for a
template. If present the option will show the svg / img.
If not it will display the name of the template.

In order to update the templates, it was decided to remove the attribute
noupdate=1 on the data file. This allows for the template to be easily
updated in the future. For this change to take effect on existing
databases, an upgrade script was created at odoo/upgrade#3169

task-2677203

Co-authored-by: Arthur Detroux <ard@odoo.com>