[FIX] website_sale: fix animation of img dragging to cart
Some templates of dynamic snippet have BS classes to manage the size of
the image (w-100, h-100, ...).
Those classes have the priority on the clone when it's created and break
the animation.
The purpose of this commit is to remove those BS classes in the clone
and let the JS function returning the correct size of the image.
It also adds to the clone the correct value of "object-fit" property to
have the perfect clone for the animation.

task-2677203