[FIX] web_editor: do not execute script from link popover widget
Before this commit, the default get method from jquery used in the
LinkPopoverWidget to update the popover icon, title and url, would use
an "intelligent guess" to determine and process the result data from the
http request. In the case of a javascript script, it would execute that
script and could lead to errors.

This commit forces the dataType to html.

task-2172311