[FIX] website: use loaded images on image wall
TO REPRODUCE:

- On empty image wall snippet with mode = 'Masonry' or 'Float', add new
images using the 'MediaDialog'.

Result: Images are then dispatched in a wrong way compared to what we
should get on this 'mode' (e.g. with 'Masonry' mode active, new images
won't be added to the "smallest-height column").

The goal of this commit is to fix this behaviour by adding "loaded"
images to the DOM before the "mode" method is applied, this way, all
image related values (size, position,...) will be computed correctly.

REMARK:

A small tweak was added to '_replaceContent()' method to prevent the
"Add Images" button removal when 'mode' updated on empty image wall.

task-2692129